requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-fontawesome==1.0
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.11.1
filelock==3.0.12
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.32
whitenoise==5.2.0
